[lldb-dap] Move the Variables struct into its own file (NFC)
Move the Variables struct out of DAP.h and into its own file to reduce
the complexity of the latter. This PR also makes the members that are
implementation details private.
